+++ CHANGELOG.md
@@ -1,3 +1,63 @@
+## [0.3.24](https://github.com/grafana/synthetic-monitoring-app/compare/0.3.23...0.3.24) (2020-09-29)
+
+
+### Bug Fixes
+
+* specify a step size for slider input ([ec2fe5c](https://github.com/grafana/synthetic-monitoring-app/commit/ec2fe5ccda7943b14c8ffacfe6a045e21a5645c0))
+
+## [0.3.23](https://github.com/grafana/synthetic-monitoring-app/compare/0.3.22...0.3.23) (2020-09-29)
+
+## [0.3.22](https://github.com/grafana/synthetic-monitoring-app/compare/0.3.21...0.3.22) (2020-09-29)
+
+## [0.3.21](https://github.com/grafana/synthetic-monitoring-app/compare/0.3.20...0.3.21) (2020-09-29)
+
+## [0.3.20](https://github.com/grafana/synthetic-monitoring-app/compare/0.3.19...0.3.20) (2020-09-29)
+
+## [0.3.19](https://github.com/grafana/synthetic-monitoring-app/compare/0.3.18...0.3.19) (2020-09-29)
+
+## [0.3.18](https://github.com/grafana/synthetic-monitoring-app/compare/0.3.17...0.3.18) (2020-09-29)
+
+## [0.3.17](https://github.com/grafana/synthetic-monitoring-app/compare/0.3.16...0.3.17) (2020-09-29)
+
+## [0.3.16](https://github.com/grafana/synthetic-monitoring-app/compare/0.3.15...0.3.16) (2020-09-29)
+
+## [0.3.15](https://github.com/grafana/synthetic-monitoring-app/compare/0.3.14...0.3.15) (2020-09-29)
+
+## [0.3.14](https://github.com/grafana/synthetic-monitoring-app/compare/0.3.13...0.3.14) (2020-09-28)
+
+## [0.3.13](https://github.com/grafana/synthetic-monitoring-app/compare/0.3.12...0.3.13) (2020-09-28)
+
+## [0.3.12](https://github.com/grafana/synthetic-monitoring-app/compare/0.3.11...0.3.12) (2020-09-28)
+
+## [0.3.11](https://github.com/grafana/synthetic-monitoring-app/compare/0.3.10...0.3.11) (2020-09-28)
+
+## [0.3.10](https://github.com/grafana/synthetic-monitoring-app/compare/0.3.9...0.3.10) (2020-09-25)
+
+## [0.3.9](https://github.com/grafana/synthetic-monitoring-app/compare/0.3.8...0.3.9) (2020-09-25)
+
+
+### Bug Fixes
+
+* stuck loading spinner on check creation ([d9a77ba](https://github.com/grafana/synthetic-monitoring-app/commit/d9a77ba12e3e355ad79eca820ed3615fd5e31534))
+
+## [0.3.8](https://github.com/grafana/synthetic-monitoring-app/compare/0.3.7...0.3.8) (2020-09-25)
+
+## [0.3.7](https://github.com/grafana/synthetic-monitoring-app/compare/0.3.6...0.3.7) (2020-09-25)
+
+## [0.3.6](https://github.com/grafana/synthetic-monitoring-app/compare/0.3.5...0.3.6) (2020-09-25)
+
+## [0.3.5](https://github.com/grafana/synthetic-monitoring-app/compare/0.3.4...0.3.5) (2020-09-25)
+
+## [0.3.4](https://github.com/grafana/synthetic-monitoring-app/compare/0.3.3...0.3.4) (2020-09-25)
+
+
+### Bug Fixes
+
+* invert noFollowRedirects in the UI ([c6d302e](https://github.com/grafana/synthetic-monitoring-app/commit/c6d302e713c4faed9c3c7c34d3560e66d72ec888))
+* make default values consistent for all check types ([56866bb](https://github.com/grafana/synthetic-monitoring-app/commit/56866bb77153ee7880f9f19b9b152db4c19a9b7c))
+
+## [0.3.3](https://github.com/grafana/synthetic-monitoring-app/compare/0.3.2...0.3.3) (2020-09-24)
+
 ## [0.3.2](https://github.com/grafana/synthetic-monitoring-app/compare/0.3.1...0.3.2) (2020-09-23)